OLI-17: Ajustes na CRUD de Helices e pedidos

+++ ollix-backend/src/Ollix.Application/UseCases/Clients/Commands/DeleteClient/DeleteClientCommandHandler.cs
@@ -29,9 +29,12 @@ namespace Ollix.Application.UseCases.Clients.Commands.DeleteClient
             var clientAppResult = await _mediator.Send(new GetClientByIdQuery(request.UserInfo, request.ClientId), cancellationToken);
             if (!clientAppResult.IsSuccess)
                 return Result.Error(clientAppResult.Errors.ToArray());
-
+           
             var clientApp = clientAppResult.Value;
-            clientApp.RegisterDomainEvent(new EntityControlEvent(request.UserInfo, EntityEnum.Client, OperationEnum.Update, clientApp));
+            if(clientApp.Id == request.UserInfo!.ClientApp!.Id && request.UserInfo.IsAdmin())
+                return Result.Error("Não é permitido excluir seu proprio cliente!");
+
+            clientApp.RegisterDomainEvent(new EntityControlEvent(request.UserInfo, EntityEnum.Client, OperationEnum.Delete, clientApp));
 
             await _repository.DeleteAsync(clientApp, cancellationToken);
